% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{beta_diversity}
\alias{beta_diversity}
\title{Compute a pairwise community phylogenetic similarity matrix}
\usage{
beta_diversity(M, branch_lengths)
}
\arguments{
\item{M}{Site-by-taxon matrix.}

\item{branch_lengths}{Vector of branch lenghts for each taxon.}
}
\value{
Pairwise matrix of sorensen similarity between sites
}
\description{
Compute a pairwise community phylogenetic similarity matrix
}
