% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_tree_ranges}
\alias{build_tree_ranges}
\title{Derive ranges for the internal nodes of a tree}
\usage{
build_tree_ranges(tree, tip_occs)
}
\arguments{
\item{tree}{Phylogeny (object of class "phylo").}

\item{tip_occs}{Site-by-taxon matrix of occurrences for terminal taxa.}
}
\value{
A site-by-taxon matrix with a column for every branch, including terminals and internal edges.
}
\description{
Derive ranges for the internal nodes of a tree
}
