% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{diversity}
\alias{diversity}
\title{Calculate spatial phyologenetic diversity metrics}
\usage{
diversity(sp, spatial = T)
}
\arguments{
\item{sp}{spatialphy object (created by \code{sphy()} or \code{simulate_sphy()}).}

\item{spatial}{Boolean: should the function return a spatial object (TRUE, default) or a vector (FALSE).}
}
\value{
A matrix or raster stack with a column or layer (respectively) for each diversity metric.
}
\description{
Calculate spatial phyologenetic diversity metrics
}
