% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tip_occs}
\alias{get_tip_occs}
\title{Occurrence matrix for terminal taxa}
\usage{
get_tip_occs(sp, spatial = F)
}
\arguments{
\item{sp}{Object of class \code{spatialphy}.}
}
\value{
A site-by-taxon matrix with a column for every terminal branch.
}
\description{
The community matrix contained in a \code{spatialphy} object includes a column for every terminal and internal branch.
This function returns the sub-matrix that includes only the terminal taxa.
}
