% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{phylo_dist}
\alias{phylo_dist}
\title{Compute a phyologenetic turnover matrix}
\usage{
phylo_dist(sp, add = T)
}
\arguments{
\item{sp}{spatialphy object.}

\item{add}{should the results be added to the spatialphy object and returned (T) or returned alone (F)?}
}
\value{
A pairwise phylogenetic difference matrix, with values of 1/PhyloSor, either on its own or as an element of \code{sp}.
}
\description{
Compute a phyologenetic turnover matrix
}
