% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snape.R
\name{snape}
\alias{snape}
\title{Smooth Neo- And Paleo-Endemism (SNAPE) analysis}
\usage{
snape(rand, palette = c("white", "blue", "red"), transform = function(x) x^2)
}
\arguments{
\item{rand}{A matrix or RasterBrick of randomization quantiles, generated by \code{sphy_rand}.}

\item{palette}{A character vector of three or more colors, corresponding to low endemism, high RPE, and low RPE, respectively.}

\item{transform}{A function that takes in a value between 0 and 1 and returns a value between 0 and 1. This
determines the shape of the color gradient transitioning from low to high endemism.}
}
\value{
A list containing two data frames. The \code{snape} data frame has a row for each community in \code{rand}, and
columns for the SNAPE color value as well as the variables used to compute it; the only variable not taken
directly from \code{rand} is \code{qPCE}, which is the parallel maximum of \code{qPE} and \code{qCE}. The \code{legend} data frame
can be used to illustrate the color scale; each row gives the color corresponding to a particular combination
of \code{qPCE} and \code{qRPE}.
}
\description{
This function implements SNAPE, an analog to CANAPE (Mishler et al. 2014) for biodiversity datasets with
continuous occurrence values representing quantities like occurrence probabilities or abundances.
}
\details{
The CANAPE method (Mishler et al. 2014) identifies areas of statistically significant concentrations of neo-
and paleo-endemism. While illuminating, CANAPE analysis requires binary-presence absence data that necessitates
thresholding continuous biodiversity data like occurrence probabilities or abundances, and it requires hard
statistical significance thresholds, both steps that
}
