% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sphy}
\alias{sphy}
\title{Create a spatial phylogeny object}
\usage{
sphy(tree, occ, spatial = NULL)
}
\arguments{
\item{tree}{Phylogeny of class "phylo".}

\item{occ}{Occurrence data across taxa and sites. Can be a matrix with a column per taxon and a row per site, or a RasterBrick/RasterStack with a layer per taxon. There must either be one taxon per edge in the tree, or one taxon per tip.}

\item{spatial}{An optional raster layer indicating site locations. Ignored if \code{occ} is a raster object.}
}
\value{
A "spatialphy" object
}
\description{
Create a spatial phylogeny object
}
