% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{sphy_canape}
\alias{sphy_canape}
\title{Binary randomization tests including CANAPE}
\usage{
sphy_canape(sp, null_model = "curveball", spatial = T, ...)
}
\arguments{
\item{sp}{spatialphy object}

\item{null_model}{see \code{canaper::cpr_rand_test}}

\item{spatial}{Boolean: should the function return a spatial object (TRUE, default) or a vector (FALSE).}

\item{...}{further arguments passed to \code{canaper::cpr_rand_test}}
}
\value{
A matrix or raster stack with a column or layer (respectively) for each metric.
}
\description{
This function is a wrapper around the \code{cpr_rand_test} function from Joel Nitta's \code{canaper} package.
In keeping with CANAPE, it requires a binary community matrix and uses a hard significance threshold (for an
alternative that utilizes continuous occurrences values and retains a significance gradient, see
\link{snape} and \link{sphy_rand}).
}
\details{
This function runs \code{canaper::cpr_rand_test}; see the help for that function for details.

It also runs \code{canaper::cpr_classify_endem} on the result, and includes the resulting classification as an additional variable, 'endem_type', in the output. 'endem_type' values 0-4 correspond to not-significant, neo, paleo, mixed, and super endemesim, respectively.
}
