% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{sphy_diversity}
\alias{sphy_diversity}
\title{Calculate spatial phyologenetic diversity metrics}
\usage{
sphy_diversity(sp, spatial = T)
}
\arguments{
\item{sp}{spatialphy object (created by \code{sphy()} or \code{simulate_sphy()}).}

\item{spatial}{Boolean: should the function return a spatial object (TRUE, default) or a vector (FALSE).}
}
\value{
A matrix or raster stack with a column or layer (respectively) for each diversity metric.
}
\description{
Calculate spatial phyologenetic diversity metrics
}
\details{
The function calculates the following metrics:
\itemize{
\item TR: Terminal richness, i.e. richness of terminal taxa (in many cases these are species)
\item CR: Clade richness, i.e. richness of taxa at all levels
\item PD: Phylogenetic diversity
\item TE: Terminal endemism, i.e. total endemic diversity of terminal taxa, aka WE
\item CE: Clade endemism, i.e. total endemic diversity of taxa at all levels
\item PE: Phylogenetic endemism
\item Em: Mean endemism (derivation is equivalent to E / CR)
\item PDm: Mean phylogenetic diversity, i.e. branch length of mean resident (derivation is equivalent to PD / CR)
\item PEm: Mean phylogenetic endemism, i.e. branch length / range size of mean resident (derivation is equivalent to PE / CR
\item BEm: Mean branch length of the endemics
}
}
