% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{sphy_prioritize}
\alias{sphy_prioritize}
\title{Phylogenetic conservation prioritization}
\usage{
sphy_prioritize(sp, protection = NULL, lambda = 1, level = 1)
}
\arguments{
\item{sp}{spatialphy object.}

\item{protection}{Starting protection status, either a raster or vector, with values between 0 and 1 representing their conservation effectiveness. If this argument is not supplied, it is assumed no existing reserves are present.}

\item{lambda}{Shape parameter for taxon conservation benefit function.}

\item{level}{Effectiveness level of proposed new reserves (number between 0 and 1, with same meaning as starting \code{protection}).}
}
\value{
A ranking of conservation priorities, with low values representing higher priorities.
}
\description{
Create a ranking of conservation priorities using greedy forward stepwise optimization. This
}
