% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{sphy_rgb}
\alias{sphy_rgb}
\title{Color space ordination of spatial phylogentic composition}
\usage{
sphy_rgb(sp, method = "nmds")
}
\arguments{
\item{sp}{spatialphy object.}

\item{method}{Ordination method, either "cmds" (classical MDS) or "nmds" (nonmetric MDS, which is slower but often preferred)}
}
\value{
A raster or matrix with layers or columns (respectively) containing RGB color values.
}
\description{
Color space ordination of spatial phylogentic composition
}
