% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sphy_simulate}
\alias{sphy_simulate}
\title{Simulate a toy spatial phylogenetic dataset}
\usage{
sphy_simulate(n_tips = 10, n_x = 20, n_y = 20, boolean_ranges = F)
}
\arguments{
\item{n_tips}{Number of terminals on phylogeny.}

\item{n_x}{Number of raster cells in x dimension of landscape.}

\item{n_y}{Number of raster cells in y dimension of landscape.}

\item{boolean_ranges}{Should simulated ranges be boolean (T) or probabilities (F, default)?}
}
\value{
A list containing a phylogenetic tree and a raster stack of terminal taxon ranges.
}
\description{
Simulate a toy spatial phylogenetic dataset
}
