% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_raster}
\alias{to_raster}
\title{Convert a site-by-variable matrix into a raster brick}
\usage{
to_raster(m, template)
}
\arguments{
\item{m}{Matrix.}

\item{template}{Raster layer, with number of cells equal to the number of rows in m.}
}
\value{
A raster brick.
}
\description{
Convert a site-by-variable matrix into a raster brick
}
