% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{sphy_dist}
\alias{sphy_dist}
\title{Compute a phyologenetic turnover matrix}
\usage{
sphy_dist(sp, normalize = TRUE, add = TRUE)
}
\arguments{
\item{sp}{spatialphy object.}

\item{normalize}{Logical indicating whether occurrence values should be divided by community totals.}

\item{add}{Logical indicating whether results should be added to the spatialphy object and returned (TRUE) or returned alone (FALSE)?}
}
\value{
A pairwise phylogenetic difference matrix, with values of 1/PhyloSor, either on its own or as an element of \code{sp}.
}
\description{
Compute a phyologenetic turnover matrix
}
