% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{sphy_diversity}
\alias{sphy_diversity}
\title{Calculate spatial phylogenetic diversity and endemism metrics}
\usage{
sphy_diversity(sp, spatial = T)
}
\arguments{
\item{sp}{spatialphy object (created by \code{sphy()} or \code{simulate_sphy()}).}

\item{spatial}{Boolean: should the function return a spatial object (TRUE, default) or a vector (FALSE).}
}
\value{
A matrix or raster stack with a column or layer (respectively) for each diversity metric.
}
\description{
Calculate spatial phylogenetic diversity and endemism metrics
}
\details{
The function calculates the following metrics:
\itemize{
\item TR: Terminal richness, i.e. richness of terminal taxa (in many cases these are species)
\item CR: Clade richness, i.e. richness of taxa at all levels (equivalent to PD on a cladogram)
\item PD: Phylogenetic diversity
\item TE: Terminal endemism, i.e. total endemism-weighted diversity of terminal taxa (a.k.a. "weighted endemism")
\item CE: Clade endemism, i.e. total endemism-weighted diversity of taxa at all levels (equivalent to PE on a cladrogram)
\item PE: Phylogenetic endemism, i.e. endemism-weighted PD
\item Em: Mean endemism (equivalent to CE / CR)
\item PDm: Mean phylogenetic diversity, i.e. branch length of mean resident (a.k.a. "RPD"; equivalent to PD / CR)
\item PEm: Mean phylogenetic endemism, i.e. branch length / range size of mean resident (equivalent to PE / CR)
\item BEm: Mean endemism-weighted branch length (a.k.a. "RPE")
}
}
