% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{sphy_rand}
\alias{sphy_rand}
\title{Randomization tests including CANAPE}
\usage{
sphy_rand(sp, null_model = "swap", spatial = T, ...)
}
\arguments{
\item{sp}{spatialphy object}

\item{null_model}{see \code{canaper::cpr_rand_test}}

\item{spatial}{Boolean: should the function return a spatial object (TRUE, default) or a vector (FALSE).}

\item{...}{further arguments passed to \code{canaper::cpr_rand_test}}
}
\value{
A matrix or raster stack with a column or layer (respectively) for each metric.
}
\description{
This function is a wrapper around the \code{cpr_rand_test} function from Joel Nitta's \code{canaper} package.
}
\details{
This function runs \code{canaper::cpr_rand_test}; see the help for that function for details.

It also runs \code{canaper::cpr_classify_endem} on the result, and includes the resulting classification as an additional variable, 'endem_type', in the output. 'endem_type' values 0-4 correspond to not-significant, neo, paleo, mixed, and super endemesim, respectively.
}
