% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{sphy_regions}
\alias{sphy_regions}
\title{Phyologenetic regionalization}
\usage{
sphy_regions(sp, k = 5, method = "hclust")
}
\arguments{
\item{sp}{spatialphy object.}

\item{k}{Number of spatial clusters to divide the region into (Positive integer).}

\item{method}{Clustering method; currently only "hclust" is supported.}
}
\value{
A raster or matrix with an integer indicating which of the \code{k} regions each site belongs to.
}
\description{
Phyologenetic regionalization
}
